version 1.1

+++ BMobileWeb/src/modelo/Cuenta.java
@@ -8,12 +8,12 @@ public class Cuenta implements Serializable{
 	 */
 	private static final long serialVersionUID = 1L;
     private String celular;
-    private double saldo;
+    private float saldo;
 
     public Cuenta (){
     }
 
-    public Cuenta(String celular, double saldo) {
+    public Cuenta(String celular, float saldo) {
         this.celular = celular;
         this.saldo = saldo;
     }
@@ -23,7 +23,7 @@ public class Cuenta implements Serializable{
         return celular;
     }
     
-    public double getSaldo() {
+    public float getSaldo() {
         return saldo;
     }
 }
